src/main/java/com/nisum/api/cl/domain/usecase/CreateUserUseCase.java
package com.nisum.api.cl.domain.usecase;

import com.nisum.api.cl.domain.model.user.User;
import com.nisum.api.cl.domain.model.user.dtos.UserDTO;
import com.nisum.api.cl.domain.model.user.gateway.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class CreateUserUseCase {

    private final UserRepository userRepository;

    public UserDTO addUser(User user){
        return userRepository.save(user);
    }
}
